Move OpenChisel wrapper into chisel_bridge/ subdirectory with own CMake
Refactor: chisel_manager.cpp/.h and thread_event.h now live in
chisel_bridge/ with a dedicated CMakeLists.txt compiled at -O3 with
aggressive flags (-ffast-math, -funroll-loops, -ftree-vectorize,
-ffp-contract=fast, -fomit-frame-pointer).

Main library stays at Debug (-O0 -g) for full debuggability of
rendering, skinning, navmesh, multiplayer, etc.
