+++ app/src/main/cpp/chisel_bridge/chisel_manager.cpp
